need to make certain that this all works, but everything is hooked up now
